requirements.txt
flask
flask-cors
Jinja2
PyYAML